Add Survey Form project about neighborhood stray animals
This project is part of the Responsive Web Design certification on freeCodeCamp. It is a survey form designed to collect feedback from residents on how to improve the quality of life for stray animals in the neighborhood.

The form includes:
- Required fields for name, email, and age
- Radio buttons, checkboxes and dropdown,  to gather user preferences
- A textarea for additional comments

All HTML elements follow best practices and meet the user stories and validation requirements set by the project.